Add a note about enabling bluetooth in Chrome
This commit adds a note about how to enable bluetooth in Chrome.

The motivation for this is that I had to enable this and others might as
well, and this will save them having to search for how to do that
(which might turn them off of trying out the game).

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>